[roll] Roll fuchsia [cpp][fidl] codegen cleanup
A very very common pattern when generating code for types that may
contain handles is to do:
  {{ if .IsResourceType }}
    {{ IfdefFuchsia }}
  {{ end }}
  // some stuff that can't build on host
  {{ if .IsResourceType }}
    {{ EndifFuchsia }}
  {{ end }}

I've added a pair of template functions so that this can be simplified
to:
  {{ IfdefFuchsiaIf .IsResourceType }}
  // some stuff that can't build on host
  {{ EndifFuchsiaIf .IsResourceType }}

The conditional is *always* .IsResourceType so this could have been
  {{ IfdefFuchsiaIfIsResourceType . }}
  // some stuff that can't build on host
  {{ EndifFuchsiaIfIsResourceType . }}
but I feel like that's a little bit too magic.

There are a couple of little other clean-ups in there that I came across
while working on it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1267024
Original-Revision: 9e7116924169910dc8283f7807e3f03578c678a2
GitOrigin-RevId: 326afc27cefee11b8d2a40da880e500919db7538
Change-Id: I63f707072da4453b4d17952859fb216c02c95521