[roll] Roll fuchsia [power-manager] Remove fake dependency
It's not longer important for power manager to outlive driver manager.
As such we can safely remove the fake dependency which was added to
ensure it does outlive it.

The lifecycle stop hook was also implemented to ensure it can exit
gracefully once it needs to stop.

Original-Fixed: 42071033
Test: component start power_manager; component stop power_manager
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259367
Original-Revision: e81e21e22346586cfa78aabe97ae8c1318fa7bb4
GitOrigin-RevId: 2a8e6b0f7f9c20708254cd7faf0cfffb533e6127
Change-Id: I2cbf64a9147d68007923707358e1919267ac976f