[roll] Roll fuchsia [kernel][arm64] Remove secondaries_released synchronization
This change removes an unnecessary synchronization variable,
secondaries_released.  The variable is unnecessary because the
variable is set to true before the secondaries are even started.

Why is it safe to remove this?  The variable was set to true during
arch_init.  The secondaries are started by platform_init and would
spin until the value is true.  However, arch_init is called before
platform_init so by the time the secondary are started, it's already
true.

Original-Bug: 414456459
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257704
Original-Revision: 0247b9756798c971efcbdecd74f5ea2d76041c57
GitOrigin-RevId: 0b2deb592bf620b17a93cb9ba51d97bb6eb61ba0
Change-Id: Ifa348c1ec098f510f6df0b10ae5f9d7c9cb27781