[roll] Roll fuchsia [ffx] Correct shutdown timeout errors.
This provides correct information when waiting for a device to shutdown.
Now if there's a timeout waiting for a device to shutdown the error
describes being unable to await the device shutting down instead of
being unable to contact the device.

Original-Fixed: 386408278

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269207
Original-Revision: f76edd89aadec3f59f9ebd7ae5fdab18625e134f
GitOrigin-RevId: 63b5c060c8cb98a88366854329e541398ae8794c
Change-Id: Ia90ace5599b43cdad06873519fe1e98e2eac93e9