[roll] Roll fuchsia [trace_manager] Don't double copy buffer in streaming mode
The tracing system doesn't differentiate between setting
"rolling_data_end" to "0" to mean no data has been written, and "0", we
haven't yet committed the buffer. This is problematic in streaming mode,
as it causes us to attempt to uncondionally read the inactive buffer
after the trace ends. This happens because after we transfer the buffer,
we "reset" the rolling_data_end to "0".

In trace manager, when we copy the buffer out in streaming mode, we now
check last_wrapped_count_, which is our local copy of the last buffer we
transferred and compare it to the tracee's wrapped count before
attempting to copy the buffer.

Original-Bug: 312790394
Multiply: fuchsia-pkg://fuchsia.com/trace_manager_integration_tests#meta/streaming_test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266450
Original-Revision: becc20eb36e12e09539481a6565b96eff024e9e9
GitOrigin-RevId: 27dc7c857dadc44a13575c4204088dc27776accc
Change-Id: Ieaf5ed807b0c7b7b4cb9aabb9ccadf5b3ed79e26