[roll] Roll fuchsia [kernel][phys] Introduce EarlyBootZbi placeholder
An upcoming change to the arm64 ZBI spec will exclude the data
cache lines corresponding the address range of a loaded data ZBI from
being guaranteed clean. This means we'll want a 'coherent' way to access
the ZBI before we re-enable the data cache. To that end we introduce the
placeholder for an "EarlyBootZbi" abstraction and plumb it through the
main signatures. A follow-on change will actually introduce a proper
implementation.

Original-Bug: 408020980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264028
Original-Revision: 847d040419131c5b8d36726263314ebfa30c213a
GitOrigin-RevId: fafa16fa3514be74f6cd1d76512563fdf2abd36c
Change-Id: Iabf08c4d72c4d2df4835bf77d76451f67dd49de8