[roll] Roll fuchsia [interconnect] Introduce a new core driver
This driver will be used as a layer between vendor interconnect drivers
and downstream clients. It takes a graph of nodes and edges and finds
the relevant paths between nodes that downstream clients will want to
set bandwidth parameters on. It aggregates requests from the various
clients and presents them to the vendor driver to take action on.

This driver was chosen to to be implemented in rust due to requirements
around heavy graph manipulation and the rust standard library generally
having an easier time of providing good tools for working with graphs
(short of using linked lists). Keeping the graph relatively flat rather
than a series of distributed linked lists was done to keep things
simpler.

Test: fx test interconnect
Original-Bug: 405206028
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261365
Original-Revision: 4a696a602be3387e943c8ac6c6e53fe591c085d8
GitOrigin-RevId: c0ee1dd00b9648e67655051eeea649269a87d843
Change-Id: Iedac5da964e8c1804447b4ede60a5f84663f2248