[roll] Roll fuchsia [ffx_trace] Filter for "kernel:ipc" when symbolizing
A non kernel trace provider could emit a flow with an id with conflicts
with an ipc flow id. This could cause an id to be contained in the
`flow_state` mapping which points to a non-step containing kernel:ipc
flow, and a step containing non kernel:ipc flow which would trigger the
assert.

This filters non "kernel:ipc" events before attempting to check for flow
stages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271144
Original-Revision: 787b28968c4e45cae3ee482c7ab38952e2494865
GitOrigin-RevId: c5394e4f554ffcd513b0a7029936b25e2bee1c0e
Change-Id: I715d1390a9bda2eec33f3a8d1e8cb4927c9bd589