[roll] Roll fuchsia [kernel] Track dropped records in ktrace per-CPU mode
Modifies the per-CPU implementation of ktrace to track the duration,
number, and size of the records it drops due to buffers being full.

All tracking is done entirely by the writers, meaning that:
1. A writer begins tracking when it fails a reservation for the first
   time.
2. A writer emits a duration to the KTrace buffer containing the
   aforementioned statistics when there is space to do so.

An alternative design was considered in which the reader would perform
step 2 as soon as it read enough data out of the trace buffer, but this
was deemed more complex, as one would need to start thinking about how
to synchronize the reader's and writer's observation of the dropped
record statistics.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262925
Original-Revision: fe713efb63dd2a9c7f1a3b3ede63b73c02e1b9d0
GitOrigin-RevId: 221c1fd22eb0cab183c47d39a6c0a4d190ef309d
Change-Id: Ic55489d82b405ca3f49d1989f5743663d66f6b82