[roll] Roll fuchsia [storage][sparse] add resparse_sparse_img
resparse_sparse_img takes a SparseReader<Read + Seek>, an output
directory and a max_download_size and generates a set of Sparse Images
for it.

This departs from previous approaches which unsparse a sparse image to a
temporary file and then build a new sparse image from the unsparsed one
as it loses the resolution of what is a Skip Chunk (which is written as
a block of 0) and what is a Fill Chunk of 0.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264027
Original-Revision: a3efaf1d3c62da619c1b32a6cfddbd55a919c798
GitOrigin-RevId: 3dc9a4510f7f2b23e2149f92a685473bb708e699
Change-Id: I82434fced51ae1b993185d257d11030f4a6e3e7a