[roll] Roll fuchsia [drivers][power] Refactor responsibilities
ManualWakeLease is now responsible for managing interaction
with SystemActivityGovernor protocols which includes:
  * Acquiring wake leases
  * Responding appropriately to suspend/resume notifications

ManualWakeLease no longer wraps a TimeoutWakeLease.

TimeoutWakeLease now wraps a ManualWakeLease.
TimeoutWakeLease is responsible for accepting a timeout
value and updating it as appropriate. A TimeoutWakeLease
uses ManualWakeLease's Start() and End() methods to
establish the time bounds during which we should claim a
wake lease, if the system is suspended, or claim a wake
lease if the system starts to suspend.

Original-Bug: 412787822

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264034
Original-Revision: 2ebca47164e9358b77d7c33e1453639fc4389689
GitOrigin-RevId: e0756df045a16c6d3a9f87673c9af95eea8c19b6
Change-Id: I4173a5313960d553e116a05d520cb5b402f381eb