[roll] Roll fuchsia [kernel][phys][userboot] Hand off decoded ELF details for userboot
This adds some machinery in physboot for handing off a PhysVmo
containing and ELF file along with loading information.  This is
now used for userboot.  So far, the kernel proper just uses this
to assert that the loading details extracted from ELF in physboot
match the built-in expectations.  Later changes will start to use
these handoff bits as the source of truth for userboot.  Changes
later yet will eventually leverage this mechanism for the vDSO.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264245
Original-Revision: 85aa9a7adfcadb6c6a789a653478da5168d18fec
GitOrigin-RevId: 31e7f879591a163972ac44142d582ad6fc701d14
Change-Id: I79a97c0a90ecc660257cf6a511be59a7c20c242d