[roll] Roll fuchsia [dash] Add power builtin
This is a migration of commands currently living under dm. The dm
command stand for devmgr, the predecessor to driver-manager which used
to do a lot more in the system. It no longer makes sense to have the
power commands live underneath it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271608
Original-Revision: 1dcc63db9fa3e58a0395f14beb8635d10ff871c6
GitOrigin-RevId: b19e229c20956e349272142b1a4bfc0d5b76625d
Change-Id: Ia3e3ce0e29200297c2f1c8ed7bb0c6c87f0d59f7