[roll] Roll fuchsia [ffx] Support "ffx -t target list"
A target list query can be specified with "ffx target list <query>", but
usually users specify a target with "-t". This CL allows "ffx -t <query>
target list" to also be used.

Note that the CL is slightly tricky, in that normally "-t" is handled
by querying the default target. But we don't want the _default_ target
to override, we only want to override if "-t" was specified on the
command line.  So we explicitly look at the FhoEnvironment copy of the
FfxCommandLine to implement the correct logic.

This introduces a behavior change: if a script were already calling
"ffx -t <target> target list" and checking for a exit status of 0,
this will break such functions. Such code should be changed to not
specify a target unless it is actually checking for the specified
target.

Original-Bug: 344903757
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269066
Original-Revision: 93f05458121efef9f7ecc1184b2b6fbc22b5881d
GitOrigin-RevId: daad1a2544747cda333df3909a0eaea2193faf5f
Change-Id: Ie2d0f917f84e25d53e253b1fabb34af7b82ce430