[roll] Roll fuchsia [sestarnix] Fail component launches without Security Context
If the Starnix container has SELinux configured then either the
container or the component must specify the Security Context under
which it should run. This removes the fall-back to inheriting the
Security Context of the init process, which is invariably not what
the component author expects.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271005
Original-Revision: 333e3a9a93a7bd6b367c14d828c7205c1731f0be
GitOrigin-RevId: 4a6f754bc90243ff943b33f2392fa0f6a71c88af
Change-Id: I88bada96b234b22a96e6747c67f306576db91d51