[roll] Roll fuchsia [magma] Support a trusted FIDL protocol
The server of CombinedDevice can now be trusted or untrusted.
At the moment the only difference this makes is that trusted
clients can have a priority higher than medium.

Update the in-tree MSD CMLs to reflect that they are have this
service.

Original-Bug: 402461734
Test: magma_sys_driver_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1258630
Original-Revision: 77cda5855f449a3effd2d7eaeffb17121f94663d
GitOrigin-RevId: 68ab7618b76e4aa9dcf91fcaa6cf5f2f1797fa46
Change-Id: I24326501cc3c4a12194d2839e6e05794299d9108