[roll] Roll infra/recipes [testing_requests] Add luci-auth linux-arm64 pin
We use this version of luci-auth on testing tasks, some of which run on
linux-arm64, so we should make sure that our pinned version of luci-auth
is available for linux-arm64.

This was causing the recipe-tools-roller to fail [0][1] because the new
version of luci-auth wasn't available for linux-arm64, presumably
because the build was broken for that platform, or due to a race
condition during uploading.

[0] https://fxrev.dev/1272027
[1] https://screenshot.googleplex.com/8Q4TQik8rtC3ZQf

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1272244
Original-Revision: f4854254a7fa341cde20222b5214e37ad2acac5f
GitOrigin-RevId: eb8f4345608ead7788966167ae23c6217dca4e88
Change-Id: I6327f1d76c9e2b9aaa4fa8f144cc073dadd717db