[roll] Roll fuchsia [kernel] Return required buffer size from zx_ktrace_read
The existing implementation of zx_ktrace_read returns the amount of
available data when the passed in user pointer is nullptr. This change
modifies the per-CPU implementation to match this behavior, with the
slight difference that it always returns the total size of all per-CPU
buffers, since that's what we require in this mode.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272209
Original-Revision: 3c6f0f1e7e4f36f4f3177580b831f90e13fedc4c
GitOrigin-RevId: 8460e9fb1f43dce6cd1fd82cded13db73ff03f36
Change-Id: I6d11594ffb795520c0227eac2f783c2ca8143377