[roll] Roll fuchsia [kernel][vm] Prevent NextChild returning root_
Corrects the TreeWalkCursor::NextChild method to avoid spuriously
returning the root_ node, specifically in the VmCowPages::Resize
operation (this spurious return is benign when performing a
RangeChangeUpdate).

In a race between dropping the current lock and acquiring the child
lock the child could be destroyed, causing it to not have a parent_. In
this case we, correctly, call UpdateCurLocked to re-establish cur_,
which is still the root_, but then just return despite not having
actually moved to a child.

This commit corrects this and will retry internally in NextChild if a
race causes the node deletion so that the stated contract of NextChild
is upheld (that it moves to a child or sibling or returns false).

Although not the cause of the specific assertion failure in b/415665459
this was found during that investigation as it causes other, non
kernel assertion triggering, errors in the Resize operation.

Original-Bug: 415665459
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271665
Original-Revision: ae576b8174d2b014e9f271b37f8749151d9981b7
GitOrigin-RevId: a071f84fa8b37bfc9a511fda1624b42a78909ed1
Change-Id: Id496755a0d33d4031199f5da48c68f0b67dbf607