[roll] Roll fuchsia [component_manager] add FIDL to get own sandbox
Adds a new FIDL API for a framework protocol that can be used to get a
component's own sandbox.

As this is not yet ready for wide consumption, the implementation of the
protocol checks that the component accessing the protocol is a builtin
component. In other words, component manager is currently the only
allowed consumer of this protocol.

Original-Fixed: 415968285
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269704
Original-Revision: 929585add9331d52a387ea69d1f7093bd84416b2
GitOrigin-RevId: 9ddccb0681ffe1faeda24907af356e0e515349c6
Change-Id: Ia013e2b53ef0d2bd8b70bc15ca99507a211937ec