[roll] Roll fuchsia [component_manager] add GetChildOutput to fuchsia.component.Realm
Adds a new function named GetChildOutput to the fuchsia.component.Realm
protocol. This function can be used to open the component output
dictionary of a child component. This dictionary has access to the same
things as the child's exposed directory, but can be interacted with
using the "bedrock" APIs.

Original-Fixed: 415965735
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269705
Original-Revision: ca19492aaa52a718d91fd2f38f9bdb08a2447415
GitOrigin-RevId: e7b0951d98572f2eb53203e596732d915246277b
Change-Id: I69fcfab1161894e694686f0a5cec87bccfa8665e