[roll] Roll fuchsia [component_manager] add built-in dispatcher component
The dispatcher component can be given configuration capabilities telling
it what capability name and type its dispatching, and the URL of a
component to dispatch to. It will expose a dynamically populated
dictionary holding the capability it dispatches. Any routing requests to
that capability will cause the dispatcher to launch a new worker
component with the given URL, and it will then forward the route request
to the worker. The worker is watched for a stop event, at which point
the worker is destroyed.

Any capabilities offered to the dispatcher are passed through to the
workers.

Original-Fixed: 415968290
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245065
Original-Revision: ad0a94f5b7395be689e0e2a0bf9881a667025285
GitOrigin-RevId: 7ee5d21599fb0d571cf6d7bd88efbf5fa578f911
Change-Id: If885aad69548c74f9af24ffb605ea08de7a809b4