[roll] Roll fuchsia [ffx] Add "is_manual: bool" to ffx TargetInfo
We'd like to add an indication to users of whether a target as
added manually or not when using ffx target list. This CL implements
one part of the puzzle -- allowing the daemon to provide whether it
is manual when a client queries the TargetInfo.

An alternative, full-featured data structure was also a possibility
(indicating the source -- discovered by mDNS, USB, etc), but the existing
daemon code is so convoluted that it would be difficult to add only
to achieve a simple goal, and since the daemon will eventually be
going away it didn't seem like a good time investment.

Original-Bug: 287693356

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271886
Original-Revision: a031e487c5c2ac98b937716a937907f00a812b0a
GitOrigin-RevId: 4b53f0ccd8792a69adeabc486a4dfe6e2bdcc647
Change-Id: Ib315dfe67ba4149768a6a257dc435542bd88d6a1