[roll] Roll fuchsia [kernel][vm] Page based slab allocator
Adds a slab allocator that uses pages as the unit of slab, and stores
its per slab metadata in the vm_page_t. This allocator is intended to
be able to be used by the other parts of the VM system itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1270505
Original-Revision: a001545b76b37e6e1a8e2031ae4b47ffa00689b3
GitOrigin-RevId: d4ffa9f33478c2e9d67c887ab5ebadb0f002ad47
Change-Id: I00551dd8cb559ef5439f43fce795f9312a2d81ee