[roll] Roll fuchsia [starnix][binder] Fix transaction dead replies
Previously we would insert `DeadReply`s into the sending thread's
command queue regardless of whether or not the reply was to the
currently active transaction of the sending thread.

This meant that the sending thread would process the `DeadReply` out of
order, and pop the wrong transaction off the stack.

This was an uncommon scenario, as it requires both a cycle in the binder
transactions between processes, and for one of the processes in the
cycle to die while multiple transactions are in flight.

Original-Bug: 414799731

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273164
Original-Revision: c2b6c066a2acc6c42f6220a44e7a3736039c7c52
GitOrigin-RevId: c7854f4fcafe7438a90ad8489816a273c80e6185
Change-Id: I14ee021cd9a9075e36a172d502c01ef8cb297b91