[roll] Roll fuchsia [kernel] Move per-CPU ktrace metadata record emission to boot CPU
Prior to this CL, the trace records containing the interned strings and
CPU koids were emitted on whatever CPU started tracing, not the boot
CPU. This led to situations in which strings were referenced before
the corresponding metadata record was seen. This worked in perfetto,
which can handle out of order records like this, but did not work
in the trace-reader userspace library, as that library needs to be
able to perform chunked reads. This CL fixes the issue by always
forcing the string refs to be emitted on the boot CPU.

Note that thread and process names are still emitted on whatever CPU
invokes ReportMetadata. This is fine because those records do not
need to be parsed before any other trace records.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272990
Original-Revision: 83cc512916fd9ae9edb339bc1dd8ca2593476e2a
GitOrigin-RevId: 9388708405699544eb74f661b39c6ba5e8be5ce1
Change-Id: I65f11f5d2523db5ce1b1f741dd18236b5261fa2a