[roll] Roll fuchsia [ffx] capture target list info in analytics
In order to understand the number of devices people generally have,
we'd like to add metrics for the number of devices, as well as the
type of query (not the query itself since that could be PII).

The event is called "ffx_target_list_devices", and it has two
pieces of information: the type of query ("first", "serial", etc),
and the number of devices returned.

Original-Fixed: 334926773

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272820
Original-Revision: 023e7d7705c5a3350e8d50c491458532a064d27f
GitOrigin-RevId: 9a68671907517ae8034f30cb6342acb9578c700b
Change-Id: I54ad69a508b5ca633920b11516af16bca8d6cf27