[roll] Roll fuchsia [sdk/test_manager] add break_on_failure
This CL adds the break_on_failure option to RunSuiteOptions. It already
appears in RunOptions, which is deprecated.

This brings the new test_manager API up-to-date with respect to the
deprecated API and is part of the effort to remove the old API.

Original-Bug: 309995479
Test: no behavior change, FIDL definition only

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272210
Original-Revision: e9db1e8926e480b0012c1cef9b5160b34cc4ceba
GitOrigin-RevId: e8911c1d3cc22263f6652e3e2a75b49ff72d1e45
Change-Id: I923b735b2cdfbd8efc846f7d84124c5fb4a2c2f5