[roll] Roll fuchsia [build][idk] Remove api_contents defined check
After https://fxrev.dev/1270044, `api_contents` is the only way to
specify the API contents, so remove a check for whether it is defined.

Also removed an unnecessary amending of visibility to allow the source
set atom to access the underlying source set because it does not.

Original-Bug: 412696706
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273892
Original-Revision: b541a68b10aa674977d0efac7e92ee2e55542006
GitOrigin-RevId: 7b00608bfd278b4713bc5f0601be5a495e1b103f
Change-Id: If1b7804a27c3c1bc4d380ad2ce27a72225372ea2