[roll] Roll fuchsia [memory] Add memory stalls to regular measurements
Add memory stalls to the measurements shown in Inspect by Memory Monitor
2. This will allow to display the stalls alongside bucket memory values,
if we so wish.

This change also removes the display of the stall rate, as it can now be
computed directly from the Inspect output. This paves the way to
repurpose the stall rate computation for Cobalt metrics.

BUG=399440175

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274026
Original-Revision: 0be3606057b1788ab33edc90f65077dfba805bca
GitOrigin-RevId: 5293499561a21c5229b4a59e7a13e46f81329f35
Change-Id: If3a4981217b61c11a829eea03cc7130f691d36aa