[roll] Roll fuchsia [wlan][policy] Logs disconnect on channel closure.
Prior to this change, the telemetry module was not notified
of disconnect when a client SME event stream terminated.
When driver recovery was the cause of this breakage, the
telemetry module would repeatedly attempt to query
statistics for a non-existent interface.

This change causes the client state machine to notify the
telemetry module that the client has disconnected on exit
which results in the telemetry module transitioning from the
Connected state into the Idle state so that interface
statistics will no longer be queried.

Test: fx test -o wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272124
Original-Revision: f2a8bdef6c0a6c2dda08e4ef3db1a9a00ec99d17
GitOrigin-RevId: ad3bc0b52acebb306f5d14d81284c267db168136
Change-Id: Iacd4a65d97aba6ced32a2a72c766e133bad4804d