[roll] Roll fuchsia [starnix][procfs] Untangle TaskDirectory from CurrentTask
Previously, the CurrentTask had knowledge of the TaskDirectory in order
to change the uid/gid for the /proc/pid directory when the credentials
of the current task changed. Now, we propagate this information without
needing knowledge of the TaskDirectory type.

To achieve this effect, we no longer cache the FsCreds object for the
TaskDirectory in the TaskDirectory struct. Instead, we recover the
FsCreds either from the underlying TaskObject itself or from the
FsNodeInfo object associated with the directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275444
Original-Revision: 80fcd50f67e55e9d518bbf68e6f628327b043435
GitOrigin-RevId: 3c7f1d0eff367a0bd5ff8951d838cfa448decafa
Change-Id: I0355ed0d128555a413dd5779db046a3a4086b9c0