[roll] Roll fuchsia [cpu_profiler] Better encapsulate component attaching
The logic to walk a component and attach to it leaked out into the
profiler_controller code. It also meant that the ComponentTargets were
having callbacks injected into them which is difficult to follow for a
reading, and also copies raw `this` pointers which is spooky.

This should help make the job to pregather all modules in the sampler
easier to reason about.

This:
- pulls the component walking logic into component.[cc,h]
- switches to using a Ref+Weak pointer so that we don't have dangling
  `this` references hanging around
- Changes component to only have access to the sampler instead of saving
  a `this` to the profiler_controller global state.

Original-Bug: 417233162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274745
Original-Revision: 744663cb8f70b900f3788ea59f5d0cc010b4703f
GitOrigin-RevId: 2792b0158e46b257e5b9efc44bca2d6bf5a80692
Change-Id: I07f5644647f8ecfbdca740913be4ed76644b4aaf