[roll] Roll fuchsia [cpu_profiler] Hold Process Modules in a Map
We'll be grabbing process modules and the beginning of the session, and
we'll want to update them at the end to grab any possibly new modules.
Switch to a map data structure keyed by build id so that it's easier to
incrementally update the newly found modules.

Original-Bug: 417233162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274746
Original-Revision: 8142dac4fa760b47dde953fcd81e846ef87dfc58
GitOrigin-RevId: 982132108bbf481b451851a9308ed4afdc6b3949
Change-Id: I375dceeb3a80959242546b287823b8878cbc7357