[roll] Roll fuchsia [bt][hfp-hf] SCO setup and code negotiation.
This commit adds support to to the peer task for awaiting incoming SCO
connections and SCO connect closes.  As part of this, it starts and stop
A2DP audio as needed. Some common functionality for SCO is moved to a
common library and some some changes are made to the AG implementation
as well.

Some other small things are in this commit as well, namely using the
strongly typed CodecId type and changing the name of
ProcedureManagedState to ProcedureManipulatedState to avoid confusion
with the ProcedureManager.

Original-Bug: b/42061084
Test: Tests in a coming commit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257844
Original-Revision: 1ec7b2c17785d81f5252ad9a9a57554c171b8b02
GitOrigin-RevId: 48a368292ca8c48401f4c927b069a59b9e470c19
Change-Id: I03a8716bfc9e251f81d7ecb820c0b7f85d5cf2e4