[roll] Roll fuchsia [usb-peripheral] Migrate to DFv2
Migrate the usb-peripheral driver to DFv2.

Make the usb-peripheral store UsbFunction instances in a vector instead
of fbl pointers. The goal is to make it clear that the driver owns the
UsbFunction instances and more easily catch uses after the driver has
stopped.

Make the driver clear usb-functions by simply removing them instead of
waiting for the usb-functions to signal to the driver that they have
been removed. Originally, this was necessary because the usb-functions
had to asynchronously remove the DFv1 device that they owned. This is no
longer the case as they instead own FIDL clients to DFv2 nodes which can
be dropped immediately.

Original-Bug: 416554395
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266884
Original-Revision: bdf9dfaf1089f2cd0c924af7e4d233b7c66d5df0
GitOrigin-RevId: 83e613cd8f4cd397ca56ffa360031e362cf06ec8
Change-Id: I7411d494dd9b02a3cc5e0b921a2361ca09dbee68