[roll] Roll fuchsia [bt][hfp-hf] Parse CIND test responses at set supported indicators
The +CIND: response to the AT+CIND? command needs to be hand parsed
since it conflicts with the response to the AT+CIND=? command.  This
adds a parser for that and uses the parsed values to set the supported
HF indicators.

Test: Tested at UPF London, fx test bt-hfp-hands-free-tests
Original-Bug: b/413733247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259066
Original-Revision: 57b1404055eab14cd9fe7291da7dbb1ac10c7338
GitOrigin-RevId: 456a2a06487ee8fb868a8afb132924dc8ffb6bc6
Change-Id: Id8a6e4111e4c7df74ae6ad164737eb662b4ab21d