[roll] Roll fuchsia [kernel] Fix ktrace per-CPU buffer init with uneven buffer size
The previous implementation assumed that the provided target buffer
size will divide cleanly among the number of CPUs in the system,
and that the resulting buffer size will be a power of two. This is
obviously not always the case.

This CL fixes the issue by dividing the given buffer size by the
number of CPUs, and then rounding the resulting value down to the
nearest power of two.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276860
Original-Revision: ddfbd0b099997e0c2391b3a1308f6b8ccbb77efc
GitOrigin-RevId: 2759e663c08569223faa434e45487d6439e56d62
Change-Id: I849ad7ed3991cb4ac3265a897440bd3d8732d42a