[roll] Roll fuchsia [kernel][vm] Introduce ParentContent markers
Adds a new type of marker, the ParentContent marker, that are used in
non-hidden nodes that do not have a root page source. The purpose of
the ParentContent marker is to allow for eliding a walk to the root in
the case where there is no content.

The insertion and cleanup of these markers adds significant overhead to
the VMO clone and close steps of micro benchmarks, but leveraging
ParentContent to elide walk ups makes up for it.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254746
Original-Revision: 29450677610f6a5f58358c9e19b6d960ddb0ae01
GitOrigin-RevId: 009d2d3dd7570ccd80cdec15c3920e7f44c9d6f1
Change-Id: I70258bbe39db18f5d78f1884dbc4f249f291ebae