[roll] Roll fuchsia [power] Increase history capture for PB and SAG
This patch increases the number of events captured
by Power Broker and System Activity Governor, to
improve diagnosability beyond ~5-10 minutes of history
(in the worst case we see in the fleet today).

With recent fuchsia.inspect.Graph optimizations, PB's
increase to 16k events occupies 82% of the VMO, which
leaves some headroom for safety and future adjustments.

SAG's increase to 4k events requires extra room
room in the VMO, and we see a steady state occupancy
of 85%. (For the "1000 leases" test, we see occupancy
of 93%.)

Original-Bug: 394674201
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1270428
Original-Revision: 28fd34589cd365c10b9b15a024c0598026ef1989
GitOrigin-RevId: be0859cf13811230e10ae628ecfb0600d576605e
Change-Id: I22e50d84092078dd733a9cca309b7682bca70045