[roll] Roll fuchsia [kernel][vm] Stack canary
Adds a stack canary, controllable via the kernel cmdline, that writes a
small canary pattern near the end of a stack to, heuristically,
determine if a high watermark of stack usage was exceeded.

Detection is performed on stack / thread deletion and so will not find
excessive stack usage on forever executing threads, or stack usage that
allocated, but did not write to, the portion of the stack where the
canary was. This limitations can be improved upon in future CLs.

Original-Bug: 417302663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272605
Original-Revision: cfa0e133fec4d252d7b35e532078ee3c109cc913
GitOrigin-RevId: 4674aa776b736e07e474c12f6a3df4420d32f528
Change-Id: Iad8aaa131a552754d4931f1608c333979472ccb0