[roll] Roll fuchsia [owners] Remove miguelfrde@google.com from tools/devshell/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/miguelfrde@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1270025
Original-Revision: 0e3b3a2e7f7ebcfe6c2df87cdfb84c2ed22d2767
GitOrigin-RevId: 5cd85bcc49537f3a50370fc3c4e89fb3321febb6
Change-Id: Id11915f31fa84a7356d98713e7dc9d4fd16a15ae