[roll] Roll fuchsia [mm2] Prevent a crash in mm2 when memory usage is incoherent
In order to produce reports of memory usage, memory_monitor2 typically
iterates through known claims and compares aggregated memory usage
with kernel reported memory usage.

In some cases, those two signals can be incoherent, for instant if a
claim is added or removed between the moment we consult the kernel and
the moment we process the claims. Before the change, memory_monitor2
would crash if the total memory used by VMOs as reported by the kernel
is lower than the total amount of memory used by VMOs as computed from
claims.

Those incoherences are expected some of the time and do not warrant
crashing the whole process. It is also probably not worth reporting
them either, because they are unlikely to be actionable.

Original-Bug: https://fxbug.dev/417933381
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277944
Original-Revision: 3cdae6c52640706ecde44c1f00513181790f06e2
GitOrigin-RevId: fa4a0739978b9c29ce30c1c72440859751586afb
Change-Id: Ie17b25c2e74061aa6656d0c411d77900271bc5b6