[roll] Roll fuchsia [netcfg/dns] Socketproxy netemul tests
Add tests to validate the new DNS server source for socketproxy-provided
DNS servers. Socketproxy DNS servers will be provided from the
DnsServerWatcher when there is no Fuchsia-provisioned default network
set. In these instances, the DNS servers will reflect the set of servers
known by the Starnix Networks currently in the StarnixNetworks registry.

The implementation of this functionality is in https://fxrev.dev/1271084
and is tested in this CL for ease of review.

Original-Fixed: 392152278
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192691
Original-Revision: 7d0e94c326cde9f9e05eaddc64463ce17586ccc5
GitOrigin-RevId: 396b9a801136b17d63eaf5c3483f4f631ba06f92
Change-Id: I6830aa4efc19333a94c555634aae35f1c36c8ff9