[roll] Roll fuchsia [sestarnix] Hooks to temporarily patch task attributes
This change introduces hooks to temporarily patch task attributes, and
restore them after performing an operation. Internally, the changes to be
performed are described by a `TaskAttrsPatch` structure, composed from
provided operations. Currently, two operations are exposed: using an
effective SID previously copied from a CurrentTask, and copying the
security state of an FsNode as an fscreate SID. These are exposed to
the outside-SEStarnix world as two independent hooks.

To ensure that the state is properly restored, we require callers to
provide a closure that we wrap in the apply/restore operations. In
contrast to providing an object with a `Drop` implementation that
performs the restoring, this ensures that the lifetime of the
temporary state cannot be mistakenly extended, by moving or
leaking. In particular, the state will be restored before returning to
user code.

Original-Bug: 413607595, 364568696, 369067922, 364568931
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275225
Original-Revision: f4bd0d49abe1c1acf711ec19f3ad3332a774a990
GitOrigin-RevId: 38075a9a687a9ba97f2b6f7ed5d59dabff4ff2bc
Change-Id: I7bc7566bd39b371ea7208e8770c104d87e83ffed