[roll] Roll fuchsia [components][moniker] Reduce memory usage
Replace Vec<ChildName> with Arc<[ChildName]> to make monikers cheaper to
clone. Looking at a heapdump of component manager shows that there are
a significant number of clones of Monikers. The ChildNames inside of
Moniker are cheap to clone but the Vec is not. Replacing the Vec with an
Arc removes the allocation from every clone. This has the downside of
making Monikers immutable but Monikers are cloned way more than they are
mutated.

This saves ~2.6MiB on android builds.

Original-Bug: 414632557
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262564
Original-Revision: cb292a12815668beb6ddb42f2b5b4f5cd34ec5a1
GitOrigin-RevId: cd032a435c84f03d07b775fcb75d2f0fe65a0301
Change-Id: I16b1b57c93d9b786a5bff843ba4d21e898bfe6d3