[roll] Roll fuchsia [super_parser] Update test image to smaller image
Update the test image such that the size is now 10MiB. It was 1 GiB
previously which was unnecessarily large. A large file will be
problematic when we add more test that opens and reads this file as
we'll run out of memory.

Also update the image to be a bit more complex such that we can be more
comprehensive in testing the parser. For example, having sizes for each
partition.

Original-Bug: 404952286
Test: fx test //src/storage/lib/super_parser:super_parser_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277544
Original-Revision: 6855df5db8cf3122312a2517c4e178fa8eb3d8ac
GitOrigin-RevId: 856caa07ccfd522559df73ac5ed55c3c1402ee5c
Change-Id: I4de4f7b114b0860e81b0b6e3ab16616c489b44a2