[roll] Roll fuchsia [gpt] Re-enable passthrough
This was inadvertently disabled during refactoring in fxrev.dev/1261608.

An integration test for this would be challenging to land, but if
storage-host was enabled on the storage benchmarks, we'd see a hit to
performance benchmarks, which should catch this.

In the meantime, another bug snuck in where we weren't validating
provided offsets.  Change this to happen when they are first provided,
rather than lazily verifying.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271864
Original-Revision: 9e71955a89717bf6de25d90b1ffbe3def69bab1a
GitOrigin-RevId: e15cc1c53b2dc23941db3993ebb62a41c97b2544
Change-Id: I4a2c69257569d8291e13c26e5bf59ed2a0ddc1b3