[roll] Roll fuchsia [dt][clock] Add fuchsia.clock.Name bind property
And specify it with the `clock-names` dt property. We probably want to
deprecate the fuchsia.clock.Function bind property as it's not really
scalable. If dt and bind were coupled more strongly we might be able to
use bind's support for enum extensions, but it would make the clock name
more verbose in dt.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277454
Original-Revision: 5468ddf5aeb9f1c25e1267301602438d83e2716f
GitOrigin-RevId: 3f9cb0a4c44281964c7d029477532c81df56fe2c
Change-Id: I0163e618be471c19b44660bf9a6652c96aae5e08