[roll] Roll fuchsia [fdio] Un-deprecate vdirent_t and remove vnattr_t
The former was planned on being removed as part of the io2 migration,
but we instead have removed the new Enumerate method to reduce the scope
of changes. This means that for the time being we will continue to
support ReadDirents which uses the vdirent_t format.

Remove vnattr_t as it was completely unused even at API level 16.

Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1279344
Original-Revision: 706c58f06934a57ee2eaba502e4dd12f5506e76e
GitOrigin-RevId: 2e87d3d024091a4aa8d5ee82a62ded96e50c8b08
Change-Id: If579f7cdcea710c69a92b87e875daf1620149229