[roll] Roll fuchsia [bt][hfp-hf] Handle unknown AT responses
Log uparseable AT responses and those which the current procedure
can't handle even though they were routed to it.

This should be removed long term and be replaced with dropping the
responsible peer connection, but currently we don't want to do that if
we hit an unimplemented procedure.

Test: N/A
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999610
Original-Revision: 21d34805c5e6f432f48ee2a89379b0cca61d2801
GitOrigin-RevId: 93b7373ac9570a72e7e740860a6d4aca1c5d21f7
Change-Id: If1e80d9b4e523bc8d8656fb3221414590e19d6d6