[roll] Roll fuchsia [bt][hfp-hf] Hang up calls.
This commit adds the hang up procedure and call module support for it.

The call module changes are fairly complex.  Determining when we can
stop waiting on a Call's Stream implementation is not simple. We need to
guarantee that we've both received a +CIEV(call = 0) from the peer and
either sent a State::Terminated to the FIDL client or gotten an error
because the client has already closed the channel.

Test: fx test bt-hfp-hands-free-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002861
Original-Revision: 54378ab677e977b4bd4ff4137972cb0c934bce1f
GitOrigin-RevId: 7409ffa071188b38d2e29ed710b8bfaeffb0f830
Change-Id: I76b1c88b3d9bd31d94516618eb1eb7446d5405d4