[roll] Roll fuchsia [fuchsia-async] Change Scope::wake_all to take an active guard
In preparation to switching ExecutionScope to use fuchsia-async's Scope
cancellation API, this changes `wake_all` so that it will take active
guards.  This will allow us to implement `ExecutionScope::shutdown` in
the way that we want.

This also introduces enough of the API to the portable runtime to
support ExecutionScope.

And lastly, we add an on_no_tasks_and_guards which is also needed.

Original-Bug: 416349680
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269644
Original-Revision: 74058324c2057eca269c7f1b81cf9fe75f5f23e7
GitOrigin-RevId: f054ee2685935e4d9db903bfac77f077c3a7c4e4
Change-Id: I96dd63cdfbb4ab1dbb5c409a5c7b95cb1297a69c