[roll] Roll fuchsia [tefmocheck] Fix check for Boot-test success string.
For the check that checks whether we find the boot-test successs string
even if the test fails, we need to add exceptStrings to exclude the
string when it's reported in error logs. fxrev.dev/1265145 added the
string to the error message from ReadUntilMatch() so we need to add an
except string for that log as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280484
Original-Revision: 25143c40d938e5df9137dbdfaef448aed26fe350
GitOrigin-RevId: 497d7618985d293911fa01769d8f0e9fe524e21a
Change-Id: I2185481b8d2bd57c95d7fcf39e34d9037d2503fc