[roll] Roll fuchsia [owners] Remove sezers@google.com from src/tests/end_to_end/idle/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/sezers@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257726
Original-Revision: 318d61768ad0b1d11d439526eec397f6683fb705
GitOrigin-RevId: 24e423350ed1144837b46c6cf58933b8f823c77b
Change-Id: Idf094ce7661bbdca985b40c1773315df96c858c6