[roll] Roll fuchsia [socket-proxy] Remove from unified binary
The socket-proxy is not being used on the clients of the unified binary,
so it should be removed. Products that desire to use the socket-proxy
can use the assembly flag to include it in the build.

Update assembly bundles and socket-proxy build targets to only build the
component/package portion of socket-proxy and not the gub portion.

Original-Fixed: 419010176
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281087
Original-Revision: 9e598e95341b6fd182003b616ded115582b3c22f
GitOrigin-RevId: 12cac2d4033f44289f8668b539ff796add895b8e
Change-Id: I4b191a393261472954bb14019ed516158923b78a