[roll] Roll infra/recipes [owners] Remove danikay@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/danikay@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1276748
Original-Revision: 9c70a0a397145fc18bbb704e8524c84a44f9b764
GitOrigin-RevId: 2523ca4030950be15660175d4ced31a14ce9aee6
Change-Id: Ie53fd308cdc158497bc1651d3e7d4b4d50d2c144