[roll] Roll fuchsia [owners] Remove danikay@google.com from src/lib/versioning/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/danikay@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276772
Original-Revision: 25f4e8811c966c31ff1b12c2678cace26b86cdb1
GitOrigin-RevId: d7050f6ab444f46ec597c210339080504f2bc53d
Change-Id: I23670107046eac98ccf4548f2d14940fc3c8cd69