[roll] Roll fuchsia [owners] Remove neelsa@google.com from sdk/fidl/fuchsia.time/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/neelsa@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276752
Original-Revision: 73f98b91118149f5a2a736c332aec6ea034da071
GitOrigin-RevId: 96e652fc6b076d1a8b8f331425ba392c947c7f68
Change-Id: I88a99508f30531c11bb39e53f051e2f6b671a7fc