[roll] Roll fuchsia [starnix] Assert that boot-to-UTC ref ticks is not zero
Assert that boot-to-UTC reference ticks is never zero when updating the
UTC transform in the starnix kernel vdso. Setting this value to zero
will cause any process that calculates UTC time using the vdso to
assert.

This is not necessarily indicative of the root cause of the issue
described in the bug, but it is a reasonable defense-in-depth measure to
catch mistakes in the calculation or propagation of this sensitive
value.

Original-Bug: b/415099415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265841
Original-Revision: 9479bff68c161112fbfb4a4ca8f38392e66c492f
GitOrigin-RevId: b1a606951ef3b82f2578871f0fa6e60a02688b9f
Change-Id: I9f11e2d8102175a40800c5f227c75dc73b3f49ff