[roll] Roll fuchsia [fuchsia-async] Fix Timer::reset
In Timer::reset, it's only safe to skip the lock if the state is
UNREGISTERED. In all other states, including the FIRED and TERMINATED
states, the lock must be taken and a slow check is required.

Without this fix, the added test would fail after about 10-20 iterators
on a release build.

Original-Fixed: 418235546
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280904
Original-Revision: 9ba831db5c3ab29f563150b026a682c57c1f7172
GitOrigin-RevId: b3d0414b7db8e4b183cbe41e7669236013183268
Change-Id: Iaf3cd9f0c64a311dc48ff97b77f42ede4c99fa27