[roll] Roll fuchsia [libc] Rework ThreadAllocator into ThreadStorage
ThreadAllocator was originally designed as a temporary object
that owns the storage only during the setup phase and does not
record all the information necessary to use that storage.  It now
seems more useful to have a long-lived ThreadStorage object that
is optimized for compactly storing all the necessary information.
For now, the ownership is moved between ThreadStorage and the
legacy C `struct pthread`.  When that's replaced by a proper C++
Thread object, it will just have a ThreadStorage member instead
to track the ownership over the whole lifetime of the thread.

Original-Bug: 342469121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275705
Original-Revision: b91e5b158a69ef469aa7d318a0c8a72d249e8413
GitOrigin-RevId: 1449c68f27b73fad2d88b24b1e467d6399d3d599
Change-Id: I1205990abf95c4d050642437eab8740bdfd00325