[roll] Roll fuchsia [starnix] Hold pid table lock while verifying tgid
Otherwise there is a race where the `Task` that has been upgraded can be
released and acquire the pid table lock before verify_tgid_for_task has
the chance to do so. The Task release then waits for the weak count to
become zero, but that won't happen since the verify_tgid_for_task thread
has a reference to it and is waiting for the pid table lock.

Original-Fixed: 418209292
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281164
Original-Revision: ebfdfdcb19a3572e5d3b651bc25de37c2be076ab
GitOrigin-RevId: 6e52a5736e1d7c97655e7755eb83631c00ea6aae
Change-Id: Ia2548f617add66b33192a2797c83662423d05283