[roll] Roll fuchsia [blobfs] Fixes handle close race for debug assert.
Outside of debug mode the response would still have been correct, but it
was via an unexpected code path. If a handle closure for the writer
hasn't been processed before the next open attempt then we followed a
code path that assumed we were still writing the blob but it had been
closed by the client.

The new case confirms that the writer is complete and the blob is
persistent, which obsoletes the previous "else" case below.

Original-Bug: 417273110
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281387
Original-Revision: 0fa0f76d21ea576236b7ddd80952ee1081aa365b
GitOrigin-RevId: 2e432f48722455cfce33df532bd5a784e62054a8
Change-Id: I1de50e377590db1ea9eaca14943dfc29978d961e