[roll] Roll fuchsia [fuchsia-async] Forbid wakers from changing lists
In theory, it was possible for a WakerEntry to change to a different
list in which case, bad things would happen.  This change prevents that
at the cost of having to always acquire the lock when the WakerEntry is
dropped.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281845
Original-Revision: 28357d4c4662bb86b5c9a5d8e63d965b347a1d3c
GitOrigin-RevId: c776f93af96e53b32d5edfe4ad9007b5b47a8876
Change-Id: I3cd97bc2c7ac4d00d2cc29252542a4141691a786