[roll] Roll fuchsia [netstack3] Correct TCP WindowSize::MAX
There was a bug in operator precedence when calculating the maximum TCP
window size. This led to the netstack artificially enforcing a lower
maximum Window Size than the RFC. This results in a panic for
TCP connections with a large Window Size (> 500 Mb).

While here, add the `clippy::precedence` lint to the Netstack3 crates.
This prevents similar bugs in the future.

Original-Bug: 419219829
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1282284
Original-Revision: d8dab31a6204716d8f960f6eca1601476f8df1d2
GitOrigin-RevId: dfd7027bba5809b6532acbfead48a7210d227a9b
Change-Id: I2df211dd35f2a354f33d17376e3c8388f26b4718