[roll] Roll fuchsia [ffx] Combine target_holders and daemon_proxy
ffx_daemon_proxy is on the critical path because target_holders
depends on it. But it is only one file, and target_holders is the
only crate that depends on it. By removing the crate, we can speed
up the build-time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281748
Original-Revision: 534895e40f0230ad2c5f6320aef7022338a7bdd1
GitOrigin-RevId: 01a54475a0fda156753a13cc99ec5748c321b9d1
Change-Id: I3bd0488fa59ac9f266f911883eab6d9efe8db872