[roll] Roll fuchsia [owners] Remove dhobsd@google.com from src/lib/fuchsia-hyper/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241791
Original-Revision: 70c983c38a4b7c4757af6b93765e6ca389a9357c
GitOrigin-RevId: 1fe6bf8439c63a531f0f794349c758ab5eec7094
Change-Id: I6d415befe94cbd93329f2b36d896dc6ed67500fd