[roll] Roll fuchsia [starnix][trace] Don't alloc DurationScopes when tracing is off
Currently, starnix's trace_duration macro checks if tracing is compiled
in, and if so, always allocates a Duration Scope, regardless of if
tracing is actually enabled or not. We should avoid this and only create
one if we are actually tracing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280149
Original-Revision: b2d7108744d2d4325fb1653779075098f202f9bd
GitOrigin-RevId: 5b49fd867a786b146cc923e90afb7e1368118b5a
Change-Id: If582f644b85526d7201a51c95d2778173c2383e5