[roll] Roll fuchsia [block] Return errors on bad alignment instead of asserting.
Change the block client to stop asserting on bad alignment arguments and
moves to returning INVALID_ARGS instead. Adds tests for the cases as
well.

Leaves FakeDevice as it is a testing application, and failing loudly is
better there, and leaving the file-backed device implementation as it is
specified to only build for host.

Original-Bug: 415893572
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1282625
Original-Revision: 48efb80481626569f2f799552b61ac5ed1892b9d
GitOrigin-RevId: 13f496c11e0a59818c783d9823b72e29ea6c322d
Change-Id: I2f46cb10bd4792f89ef0f89210b08445adbc3fd2