[roll] Roll fuchsia [fuchsia-async] Support half-closed zircon sockets
Add support for waking up wakers when a zircon socket is half-closed on
fuchsia_async::Socket. The existing implementation would haphazardly
return BAD_STATE if it manages to observe the half-closed state, or pend
forever if it doesn't. Given no systems currently reopen half-closed
sockets the default behavior of loudly bailing out of waiting sounds
like the right one.

Add an escape hatch for a socket that ignores half-closed, and pends
until either readable or writable are available. The plan is to use this
version in overnet (pending addition of emulated support for this) so it
can proxy the half-closed signals.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1280486
Original-Revision: 2c5ac8d2647e63e1ce28f48024fce0ad4856169a
GitOrigin-RevId: b72c1d4984f1aa5591d0e8f8171261988bb1cfdd
Change-Id: Ide39631564e09c59f4f3d7dfcf937f6a929e67db