[roll] Roll fuchsia [pkg] Merge full-resolver with pkg-resolver
In Fuchsia, component resolution is composed of two layers: the
component resolver, and the package resolver. Both Component Framework
and Software Delivery define their own resolver abstraction.

Currently, there exists an independent component (full-resolver.cm) that
adapts the package resolver to the component resolver ABI. However, the
implementation of this component is quite simple and there isn't much of
a compelling reason to keep it separate. For that reason, this commit merges
full-resolver with pkg-resolver, allowing us to save a bit of memory.

The full-resolver implementation is not changed in the commit. As a
followup, we can simply the implementation to call the pkg-resolver rust
routines directly, instead of through fuchsia.pkg.PackageResolver
protocol (which now just loops back to the same component).

 ## Resource footprint

This change saves memory on eng builds by removing full-resolver from
core. Also, the instance of full-resolver in bootstrap for driver
loading is replaced with loopback-resolver, which is considerably smaller
(it's a simple C++ program). In non-eng builds, at worst case this adds
a negligible amount of memory footprint to pkg-resolver, which now has
the code and data that previously belonged to full-resolver. If the full
resolver service is never used, the data cost is zero.

Original-Bug: 417525026

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274091
Original-Revision: c21a7bc207adccbad48f1289ded6194ee0509617
GitOrigin-RevId: 52f68fb88861d541b80a096ef4592e424376bf56
Change-Id: Ia05f8615ad95214ce22cecc8eb60726826427275