[roll] Roll fuchsia [userboot] Wait for bootstrap channel to become readable
The userboot code never waited for its channel to become
readable.  In the past, the kernel stuffed the message into the
channel before starting the process, so this was overlooked for
years.  Now the process starts running in parallel with the
kernel stuffing the bootstrap message, so it became a race.  The
race is probably an optimization so the userboot startup work
before the channel read starts earlier.  But userboot needs to
wait in case it got there first.

Original-Bug: 416531428
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283404
Original-Revision: 3a9f7cad765f0b765164c2adcbf38f56dab845a6
GitOrigin-RevId: 4a249d4501566fe205b1bce1147622abae871dfd
Change-Id: I3ea27ea0bf0e5f7ca3476ba833963c3476084698