[roll] Roll fuchsia [testing] test attribute on host no longer initializing logging
The docs for the macro #[fuchsia::test] state that when implemented on
the host, they do not initialize a logger and that the host must have
logging initialized. That was not true until this CL.

This adds tests to ensure that when a #[fuchsia::test] is run on the
host that a logger is not initialized, and that when it is run on the
target that it is initialized.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283246
Original-Revision: b6280f56fb4c51f3031a4247d35ee9a6c4063167
GitOrigin-RevId: 47e97d41a4dbc36b5f1fe503818abf191dd72cc4
Change-Id: Ida328cdf4c5f7f10762e86f5a9a03e63f8a5774e