[roll] Roll infra/recipes [cache] Don't delete caches synchronously after failure
When cleaning up the cache after a failure that would usually cause the
builder to fail, don't actually delete the files after moving them into
the cleanup directory. The cleanup directory is guaranteed to be deleted
after the Swarming task completes, so it's only necessary to explicitly
delete the files if we want to save disk space. That may be the case if we
clean the cache up at the beginning of the build when encountering a
dirty cache, but it's not the case after an error when the builder is
about to exit.

In that case it's more important to get quick feedback to CL authors
about the failure, so it's best to not delete the files, since that can
take O(minutes).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1283904
Original-Revision: b220139fe165e6a718dfbf9e853293ce390748de
GitOrigin-RevId: 68968f66c0fde40e9c0819d67c64be6d99a92f84
Change-Id: If79d9c02b6b2c00f38f6f75ddc98b816755bbb08