[roll] Roll fuchsia [drivers][power] Change meaning of return value
Change the meaning of the boolean to be more intuitive. Change
meaning of return from `TimeoutWakelease::AcquireWakeLease` and
`ManualWakeLease::Start` such that a `true` means we have a
system wake lease, if required, and false means we require one,
but don't because we didn't have one previously and failed to
acquire one now.

Original-Fixed: 418266694
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281607
Original-Revision: 14dbd723b8918135eef4822912221136d46ae84b
GitOrigin-RevId: a6a9729b686a2d23b0b4d981d39b03153303b513
Change-Id: Ie5e258c48b6e5d74b62b87486a7f783bd7ed9f1c