[roll] Roll infra/recipes [fuchsia/fuchsia] Output property to show experiments from test spec.
The spec properties are versioned properties and aren't included in the
buildbucket input properties of a build for querying. In case we want to
set an experiment on only a subset of builders, we can use this output
property to find the builds where the experiment is set to run at a
percentage even if it's not enabled on the particular build.

Original-Bug: 418256599
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1283827
Original-Revision: e77186a8e7e750ca27b35895cdd9934dfcbb0c61
GitOrigin-RevId: 449665f99f184f7ecddd82f533c6b29add183755
Change-Id: Ida6e639281c67f9a841567ab0f3b2bf8017aaa49