[roll] Roll antlion Use 20 MHz channels in ChannelSwitchTest.
Antlion has support for performing a channel switch, but current support
is limited to BSS configurations with 20 MHz channels. This limitation
appears to be the actual cause of almost all of the expected test
failures in this suite (with the exception of failures due to a vendor
driver problem tracked in b/395972717). This change forces the AP to use
20 MHz channels, so channel switch commands sent via Antlion will be
correctly issued.

The underlying AP software (hostapd, in this case) can support channel
switches for wider channel bandwidths, but additional logic in
antlion/packages/antlion/controller/ap_lib would be needed to calculate
the correct center frequency and bandwidth.

Original-Fixed: 42165602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1275505
Original-Revision: e02c9dfd065b7b7a29429216ce970be05781d2a3
GitOrigin-RevId: 07c35ca4355ca933604ca11bd23a5a8232c44324
Change-Id: Ib9387be28f82a14ada1e7849a4321864c8e9294e