[roll] Roll infra/recipes [fuchsia_roller] Fix ordering of logs
This fixes a bug introduced in
commit 1fdf1df577ac9d565cd331b380e13d087bfa751a whereby rollers would
squash multi-commit rolls in the wrong order. `multi-repo-roll-resolver`
outputs rolls in chronological order, but `api.git_roll_util` expects
logs to be in reverse chronological order, which is the order that the
Gitiles API uses.

This was causing the roller to only roll one commit at a time (since it
would apply each roll in order but end with the earliest one), but still
report rolling multiple commits in the commit message.

Original-Bug: 418290894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1289324
Original-Revision: 057cab2f44ad05f4718116eb25bfa1b399c6d67e
GitOrigin-RevId: 965af162b27aebfd77b6d76783fb36e39ebaf0c9
Change-Id: Ie7401712c3ad96ad15454e15023e291427c422ba