[roll] Roll infra/recipes [release] Correctly handle Jiri imports
commit efb8df58b7efc21182cbfe94991f4bae36c3e2ee updated this recipe to
handle fuchsia.git as an import, but it still assumed that the
`gerrithost` field was set on Jiri imports in manifests, which is not
the case. Fall back to computing the gerrit host from the remote if the
gerrithost field is unset.

Original-Bug: 385378627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1290765
Original-Revision: 8c5000bcd94d0ecb23dbcc3f028811a865d556c0
GitOrigin-RevId: 6199c7bfbf89fb896772619b28e64b82b27c56ff
Change-Id: I0d60e740b1159bed14ac146f08abef12be706364