[roll] Roll infra/recipes [rust_toolchain] Apply enzyme case workaround on stage0 build only
We are seeing checkout errors on non stage0 build on Windows and
mac. The workaround should have been a no-op for non stage0 builders
but it appears it is not. This patch limit the workaround on stage0
builders only to mitigate the failure.

Original-Bug: 420675659
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1294064
Original-Revision: 05cfc086017691f51b600be5f55c847cb999c2fb
GitOrigin-RevId: e8f9850519f6cead0bb74314822e26595efabfbc
Change-Id: I92a56e814fdb325f2c0d74ef3f381d601502120f