[roll] Roll infra/recipes [auto_roller] Summarize tryjob failures
Summarize the pass/fail status of tryjobs when processing rolls.
Specifically, list all failed tryjobs and a subset of running tryjobs,
repeatedly as status is evaluated, and, if the roll failed, at the end.

Tryjob status will be kept internal to the build, so any hidden tryjobs
that are visible to the roller task service account will only be visible
to users that have access to view that build.

Original-Bug: b/399173283
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1297027
Original-Revision: 48c8f5ab655f8d4aa39815149c1a51cf047b6fb7
GitOrigin-RevId: 7d245cf279b233b188278183d4df8507528ce978
Change-Id: Idd8d84465cf18d7ebf5baf2ab2bbb2b5a644985a