[roll] Roll infra/recipes [cherry_pick] Use integration repo if remote matches
This is a quick fix for failures like this:
https://screenshot.googleplex.com/4KZyzSGkoffcojD

That happens when trying to cherry-pick a smart-integration CL onto a
non-"sd" branch, in this case "canary". We use the branch name to
determine which integration repo to cherry-pick the change into, so if
the branch name doesn't contain "sd" then we'll fall back to using GI,
which is incorrect.

As a workaround, use an integration repo for the checkout if it matches
the remote of the incoming CL. This doesn't handle non-integration repos
like smart.git, but cherry-picks in smart.git on non-milestone branches
should be rare compared to cherry-picks in integration repos.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1304486
Original-Revision: 2c2c500c5ee3292e78f5ba980973be9bdb25a6d1
GitOrigin-RevId: 197ce2524d34214f0e861fbec7e2557a1ba1624b
Change-Id: Ia161c8f1126bd1418fb5b751f2d6eaef85271c8d