[roll] Roll infra/recipes [checkout] Avoid double-patching atomic CLs
Previously we were patching the top-level CL twice because it would be
added to its own deps once and then again during memoization:
https://screenshot.googleplex.com/7sPrLF6jEGKzTFK

I believe this extra patch is a no-op, but it's unnecessary and
confusing.

Now we only add each CL to its deps once.

Tested with led for a comparison:
https://screenshot.googleplex.com/6bv3Y5jgKrwS5xL

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1304608
Original-Revision: ea0175898b91010691293cd7d486beda8386c7ec
GitOrigin-RevId: ea8e8968e2c1fd17ef2cc0b0512f7b5df55cbf0f
Change-Id: Ie1e1b9300811f31f349a3e049d280bd4498d203e