[roll] Roll infra/recipes [assembly] Upload board / bib_sets artifacts using cipd_name field.
In order to generate unique artifact names, the existing "name" field
will be slightly modified to be unique (prefix with the container type).

A new field named "cipd_name" will contain the current value of the name
field, for backwards compatibility.

Also slightly changed the behavior of the pylint shac check, as it
was failing for me locally.

Note: this depends on fxr/1295088

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1283810
Original-Revision: 02bccd421fe0132f43576113d65980b79d06d5da
GitOrigin-RevId: bb1ea8239124daa4952fb60160ccaf7c76164f52
Change-Id: Ibc81ab7cd85f5c3c8a793fe5a9bc14fda8ab60fe