[roll] Roll infra/recipes [testing] Only read logs to present.
We only present a few logs that we expect the test results to have such
as the swarming log, syslog and serial log. We used to present all
output files and thus try to read the contents so that we could set them
in the step logs, but we no longer need to do that since we don't
present the other output files anymore. Also, this will prevent
processing failures like in the attached bug where the outputfile is an
unknown format that we can't read (in this case, it was a symlink to a
directory).

Original-Bug: 426615066
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1335642
Original-Revision: 3297aa1daee91371971af423334d475a9778f6c8
GitOrigin-RevId: 2cee2aa0f77d6479afbe26e4fe890346cf114146
Change-Id: Ia94f704276bbaf52b4ce0bb3bcbb591fe2dbe485